PrePro name, filter, trial changes
Changed the name of the docx file so it is more clear what step of the process it is. Changed the low-pass filter because it was too steep and was breaking. First four triggers have to do with when participants open and close their eyes before the videos start playing, so they are excluded from the preprocessing now.